+++ test/fixtures/c32_fragments/demographics.xml
@@ -25,7 +25,7 @@
          <id extension="24602" root="SomeClinicalOrganizationOID"
             assigningAuthorityName="Some Clinical Organization Name" />
          <telecom use="HP" value="tel:+1(312)555-1234" />
-         <addr>
+         <addr use="HP">
             <streetAddressLine>1600 Rockville Pike</streetAddressLine>
             <city>Rockville</city>
             <state>MD</state>
